requirements.txt
flask
adafruit-circuitpython-dht
RPi.GPIO
